Networking started
Okay so the local messaging over UDP is working, it is sending messages and receiving them and then sending them back with no problems. Client get's initialised with a random port number while the server's port remains fixed.
We need to figure out a way to work out hole punching, and them maybe make it work over TCP. Well, i dont really think we would need TCP instead of UDP at this point. Although there is a risk of dropping packets involved, which could, potentially result in a loss of data. But we can, theoretically, figure out a way to validate the data coming through UDP and doNOT accept the data until it is the correct format.
Another workaround would be to JUST notify the other device to pull the game file from the server when we tell it to. This essentially works as an alternative to SSE but could really work practically as the hit limit on our website would probably be far from exceeded (50,000 hits are a lot). And also this is not polling yayyayayayay.
Todos: Figure out hole punching lmao.